blog/server/requirements.txt
flask==2.2.3
flask-cors==3.0.10
mysql-connector-python==8.0.32
python-dotenv==1.0.0
